Added Revy and Caleb to "The WIP Table" (but not character select)
A non-final version of Revy and Caleb can now be accessed via the
"Select Group" button but NOT the character select screen. Table 14 is
temporarily "The WIP Table: Revy and Caleb". Also included on the table
are Samus and Shantae, as Revy has several lines targeting each of these
characters. Any character can be swapped out after the table is selected
but before the game starts.
This commit also includes small text fixes for Revy.