Corrected typos, updated Rosalina's behavoiur.xml
Rosalina's behaviour.xml had edits directly to it that were not made to
rosalina_quotes.txt. I did my best to preserve these as best as I could,
but it took over an hour just to confirm a few short lines.

Please, if you edit the behaviour.xml directly for any character, edit
the quotes file at the same time!