Update released opponents' meta.xml w/release tag
All released characters should now have a <release> tag indicating the
relative order in which they were officially released in the game.
WIP characters in the repo do not have this tag, but it should be added
when they are officially completed.
Release numbers were automatically added by an external script, which also
seems to have converted some tabs to spaces and moved some comments, but
there should be no change to the meta.xml data (other than the new <release>
tag) otherwise.