Update search logic to handle sorting modes
Added variables to keep track of the user-selected sorting mode, as well
as the sorting callback to use. As the sorting dropdown is currently in
the search menu, a sort will only take place when the Search button is
pressed.

Sorting options are stored in an object map defined in sortingOptionsMap,
which should allow more sorting modes to be added if desired, as long as
such a mode can be done programmatically.