Deal cards at a constant speed
Previously, cards would take a constant amount of time to get to their destination. This commit changes the animation to be at a constant speed. Cards arrive at nearby destinations quickly, and take longer to get to farther destinations.

The card speed is still based on the "Deal Animation" setting, but only the farthest-traveling cards will take that amount of time.