Trigger line counting on selection page change
The call to behaviour.xml to compute unique line counts is now triggered
for characters currently on the selection screen in the following situations:

(1) That character's line or image count is not already loaded, AND
(2) The user presses the Credits button, OR
    the user changes pages
On first run, this will increase the number of calls to characters'
behaviour files by four, but it should not be needed upon subsequent
loads. Ideally the calls should not be made unless the Credits screen is
displayed, so that might be an avenue worth exploring.

This commit also introduces JS Promises into the code, which are cleaner
than chaining AJAX success callbacks. However, they are newer, so they may
or may not impact older or mobile browsers (MDN suggests they are at least
supported by most). Should they cause problems, either they will need to
be rewritten, or more preferably, be substituted with a polyfill.