Differentiate individual and group counting
Counting handlers will now trigger for group members on the group
selecrion screen on page change or credits button.

Once again, if line/image counts for a character have already been loaded,
they will not be loaded again. However, because of the way the
loadGroupMemberMeta code is written, characters whose counts were loaded
on the individual select screen are different from those on the group
selection screen, so browsing groups for the first time this session WILL
cause the behaviour.xml file to be fetched.

TL;DR: Group Select screen Misty and Individual Select screen Misty are
are two different players, so the Misty behaviour.file is fetched twice.
The way group members are loaded would have to be rewritten to reduce the
number of calls further.