The Extreme Exhibitionist Update!
All content by DrankeyKrang

- New character: Mettaton (694 unique lines plus epilogue)
- New character: Moon (3059 unique lines plus epilogue)
- New character: Captain Falcon (1362 unique lines plus epilogue)
- New character: Gay Spaghetti Chef (303 unique lines plus epilogue)
- New epilogues: Rosalina
- New epilogue: Mercy
- New epilogue: Samus
- New background: Street ("Inner City Spectacle")
- New background: Green Screen ("Live in Five")
- New background: Arcade ("Ace's Arcade")

A successful Pipeline is required for this to be pushed to the online
version of the game. That might be this update or any subsequent patch.
We do not have direct control over this. This update is now fully
available in the offline version.

Thank you, DrankeyKrang!