Fixed a Zone-tan image
Due to initial problems with generating Zone-tan's images, 5-horny1 was
accidentally a duplicate of 5-horny3. This is because the 5-horny1 code
had an oddity that would close Zone-tan up unexpectedly, and it ended up
being commented out of the final set of images. However, because the
setup image had the same name but the appearance of 5-horny3, this was
overlooked for a couple of months.

I have now used KKL v47 to correct this.