Moved Willow to the offline testing folder
As space for the hosted folders is at a premium, I have moved Willow to
the offline testing folder in favor of Sayla, whose images have a very
similar combined size.

If you feel that Willow was the wrong choice, feel free to reverse this
decision. However, I based it on:
- Smallest number of unique lines: 173
- Author's last activity on the character (and Willow's files last
edited): 10 weeks ago
- Author indicated no progress made: 5 weeks ago
- Still 6 broken images despite numerous broken image reports

All other candidates have hundreds more lines and/or much more recent
development.

There can definitely be a place for Willow in the game if more work is
done on her. However, Sayla feels a lot closer to finished and will
likely enter the main roster first.