Moon image update by Horsecatdraws
Moon's orginal images are by DrankeyKrang.

As Kisekae is limited in how well it can display the kinds of things we
want to show in SPNatI, Horsecatdraws has modified some of Moon's images
to better demonstrate Moon's disrobing. The majority of the changed
images are in stages where Moon is making an impression on her shirt,
but he has also drawn her lifted shirt and lowered shorts for several
images too.