Rebuilt Alice's behaviour.xml
Alice's behaviour.xml was hand-edited and subsequently becamne an
invalid xml. For example, the targetStage values were not in quotation
marks, and I found a superfluous </case> in stage 2.

I took the time to update her source text document, which was
significantly out of date, then used this to regenerate her
behaviour.xml. I took pains to make sure that no content was lost,
including targeted lines. To the player, the only changes should be some
corrected punctuation and capitalization, as well as a small number of
stage-generic lines that were previously overridden at all stages.