make_xml.py now converts italics and ampersands automatically
If you can make my code more elegant, please do so.

This does work with specifying a new name for behaviour.xml as an
additional argument as per how_to_make_a_character.txt in case anyone
does that. However, specifying a new name for meta.xml doesn't appear to
work, even before the changes in this commit. It's possible that nobody
has even needed to use this.