make_xml.py now supports multiple start lines
Credit: Joseph Kantel

With this change, multiple start lines are entered into the
behaviour.xml in reverse order, so put your character select start line
at the end. (The second, random line it draws from this set is
unaffected by sequence.)
Also adjusted the "Straight" capitalization correction, as "straight" is
also a character tag that we don't want to capitalize.