Rebuilt Velma's behaviour.xml
This older character had her variables in a different sequence, which
typically makes it very difficult to make sure that all lines are
preserved. However, I was able to convert her old behaviour.xml to a
temporary text file using the Character Editor and generate an xml from
that to compare to the adjusted text file.

Besides reordering the lines, this rebuild:
- Fixed some broken images, as Velma doesn't have "won" images in her
last two stages
- Expanded her one targeted line to all stages
- Made her "American" tag lower case