Moved Velma to the offline_opponents folder
- At the time of removal, Velma had only 161 unique lines, not meeting
the 180 requirement to be put on the testing tables or the 200
requirement to get on the main roster.
- Like all offline_opponents, Velma is welcome to rejoin the game
through
the same process that a new character with 180+ lines follows: testing
tables, sponsorship, readmission to main roster.
- In the previous commit, I ensured that Velma's source text document
and xml were syncronized to make future improvement easier.
- As there only two Scooby-Doo characters on the main roster now, I have
removed the Mystery Inc table. On the hairy table, Velma has been
replaced by Daisy unless someone can suggest a complete character that's
even more hairy.