Moved Fred to the offline_opponents folder
- At the time of removal, Fred had only 169 unique lines, not meeting
the 180 requirement to be put on the testing tables or the 200
requirement to get on the main roster.
- Like all offline_opponents, Fred is welcome to rejoin the game through
the same process that a new character with 180+ lines follows: testing
tables, sponsorship, readmission to main roster. As we have so few male
characters, extra attention for Fred here would be particularly
appreciated!
- Fred's source text document and xml were previously syncronized to
make future improvement easier.
- As Fred was the final non-silent character with fewer than 180 lines,
a few unpopular characters are likely next to be moved to the
offline_opponents folder before we look at sub-200 characters.