Added version stamp to make_xml.py
The main game is currently on version 11.45.0, so I chose 1.45 for
make_xml.py to match it and reflect that it has been updated
significantly several times since its first iteration. However, these
numbers do not need to synchronize in the future.

I propose upping the version number here each time functional changes
are made so that we can get an accurate idea of how old an author's
make_xml is when they need tech support.