Corrected umlauts in Rinkah's dialogue
Also corrected ~player~ to ~name~ in the epilogue, and made the Flame
Tribe capitalization consistent.

As Rinkah only has an epilogue for males, female players might finish a
game to get her epilogue and be disappointed. I know this has happened
previously for Palutena (epilogue now available for both genders), and
it happened for me with Rinkah in testing.

We still don't have a confirmed reason for the special apostrophes and
ellipses being turned into umlauts, but the Character Editor is the
primary suspect.