Corrected artist credit for Zelda
MidiSnow has approached me to inform us that our Zelda model was created
by him. He has a dA post that predates Myterritory20 by four month. It
appears that Myterritory20 merely forwarded on the code for Zelda and
did not create it himself/herself. MidiSnow has approved sharing credit
with humdumthrow, who did the emotion/reaction posing for this model.

References:
-
https://midifang.deviantart.com/art/Princess-Squad-Cosplay-Contest-540292067
-
https://myterritory20.deviantart.com/art/My-Princess-Zelda-TP-563930088

The second image has been updated with the face we're using in the game,
but the image is still automatically timestamped for October 2015 .