A few fixes
Deleted the edit.bak files and gave back Zone-tan's release number.

I generated Shego's behaviour file using make_xml.py, but to my surprise, her .txt file doesn't include her markers, so they would be lost. For that reason, I didn't change her behaviour file.

This is also a test to see if the github client works. It looks incredibly useful! Seeing all the changed files with a comparison when you click on the file is gonna be of much help in the future.