Added basic female epilogue for Rinkah
Rinkah was given an epilogue badge but did not have an epilogue available for both genders. I've given her a basic female epilogue using 60% of the male epilogue's images. This could be improved greatly with some female-specific images.

I also adjusted the behaviour source text document to match the xml, adding the female epilogue and making minor corrections. Note that Rinah's markers are now in lowercase.