Added lost lines to Penny
I realize that Penny is an offline-only character, but her xml error was keeping her from loading in the Character Editor, and that means she wasn't a possible character to see pop up randomly in the Dialogue Tester. This made me sad.

While fixing her simple error (there was an extra </case> in 6-stripped), I unearthed some stage-generic lines to use to fill in her overly-repetitive dialogue. She'll never be a real contender for re-entry without a significant dialogue improvement, but hopefully this will benefit someone playing at home as they dig through our history for opponents to challenge.

This commit also deletes orig_quotes.txt, which is the same as quotes.txt but with worse punctuation and spacing.