Annie fixes
All changes and additions by are FrostFireFive.

A message to whoever generated Annie's xml the time before this commit: Please update your version of the Character Editor to version 1.15. Several fixes have been made to it, including correcting syntax for manual priority. This kind of information is in danger of being discarded if you edit an xml generated by a later version of this program.