Replaced asuna.txt with Character Editor export
In the process of revising Asuna's handwritten xml, I am getting her text document synchronized with modern times. This commit overwrites her original text document with an unrevised text document export from the Character Editor that is based on her xml. As you can see, she had many xml-only changes by her author, Dilletante.

In the next commit, I will make corrections to this document and generate a shiny new xml for Asuna. Because the whitespace and ordering are different, the diff will be worthless for the xml, so check these two diffs on the text document to see what actually changed.