Rebuilt Asuna's xmls
As per the previous commit, I rebuilt Asuna's xml files to replace the handwritten versions. Check the changes in the asuna.txt diff. You can now also get a quick overview of her targeted lines there.

While I was there, I made a very small number of corrections and additions.

A few shortcomings of the text export of the Character Editor:
- Release number forgotten
- ".png" added to selection screen image, leading to ".png.png"
- The following is added above the stage-generic lines: "stage-specific lines that override the stage-generic ones"
- Labels missing from negative stage numbers
- Targeted lines are shuffled into their stages, but having them isolated is better due to how often they are identical across stages
- Minor: self-stripping stages are grouped by number, not the traditional way

A few highlights of the text export from the Character Editor:
- Epilogue support!
- Marker support!
- Mulitple start line support!
- Targeted line support!
- Adds in templated instructions for those that need them!