Added D.Va's Cruiser to the new 'reskinned_opponents' directory
YuriSpacePirate has made a complete reskin of D.Va to recreate her 50s American skin in Overwatch. This optionally replaces all images and slightly adjusts several lines of dialogue.

The biggest challenge here is ensuring that the dialogue does not get out of sync with her primary version's ongoing improvements. To that end, I have incorporated her new dialogue in her existing dialogue file, but with a commented-out tag that can be enabled to allow her Cruiser lines to fire. As her clothing names and start lines also differ, the Cruiser alternatives are likewise commented out ready for the next time her xml is generated. See d.va_behaviour.txt for more details.

I have put the new images and new behaviour.xml in the newly-created 'reskinned_opponents' directory. For instructions on how to play against her, check the readme.txt. The file management required is much the same as playing against an offline or incomplete opponent, except the files identical to both versions of D.Va have not been duplicated here, so it's best to make a second copy of the "d.va" folder in "opponents" and paste these changed files over the old ones.

This is new territory for the game, so please let me know if you encounter any difficulties.