Retired Tracer to offline_opponents
As she has the fewest unique lines for a non-silent female character (193), Tracer has been moved off the main roster and into the offline_opponents folder. Should a contributor wish to mount an attempt to revive Tracer, they are welcome to make improvements and then submit her for the testing tables where she will be subject to the same sponshorship standards as any character wishing to enter the main roster.