Retired Shazza and Sola to offline_opponents
As Shazza, Sola, and Sheena fared worst on a recent popularity poll, these are earmarked for retiredment. The first two of these are moving to offline_opponents in this commit. Should a contributor wish to mount an attempt to revive either of them, they are welcome to make improvements and then submit her for the testing tables where she will be subject to the same sponsorship standards as any character wishing to enter the main roster.