Retired Sheena to offline_opponents
As Sheena fared third-worst on a recent popularity poll, she is moving to offline_opponents in this commit. Should a contributor wish to mount an attempt to revive Sheena, they are welcome to make improvements and then submit her for the testing tables where she will be subject to the same sponsorship standards as any character wishing to enter the main roster.