Removed some hardcoded references to the opponents folder
What this is:
    loadOpponentMeta(OPP + folder, oppDefaultIndex);
The above line and four like it in spniSelect.js were a hardcoded way of looking for every opponent in the opponents folder. (In spniCore.js, OPP is defined as "opponents/".) Because I've removed this code shortcut, each entry in listing.xml now needs the folder name specified.

Why this is:
Not every character lives in the opponents folder. Some characters live in folders off the saves folder instead. By requiring the folder name to be listed, we can name other folders including saves/offline_opponents.

Advantages:
Last week, Joseph Kantel made a helpful bug fix that allows the selection screen to display characters even if some folders are missing or misspelled. Coupled with this change, we can now list characters that may or may not be present in the version of the game you're playing. For example, spnati.net does not have any characters from the saves folder, as the saves folder is not hosted in the online version. However, your offline copy of the game *does* have this file, so this character would appear for you offline. Additionally, a couple of web sites rehost the entire game, saves folder and all. Effectively, they are hosting the "offline" version. Listing saves folder characters will allow these characters to appear in these rehosted versions. Because we don't want the main roster to be crowded with low-quality and incomplete characters, I invite discussion on to what extent we'll use this. For example, an Overwatch table including Tracer (currently retired) would be possible without adding Tracer back to the spnati.net version, with Tracer being simply missing from the group in the spnati.net version.

Limitations:
The first limitation is smaller. Currently, a missing character in a group displays as a broken image. Better handling of this is possible.

The second limitation affects gameplay.
    target = "opponents/" + target + "/";
Targeted lines are hardcoded in spniBehaviour.xml to look for the opponents folder + the targeted folder name. Currently, characters that aren't in the opponents folder cannot be targeted. Fixing this is not as easy, as removing the hardcoding here would require adding the folder path to every targeted line in the game. Further, it would be incompatible with the way that the Character Editor reads and creates files. If anyone knows the solution to this one, please lend us your assistance.

Request for bug reports:
I have tested the game in all aspects, and this change does not appear to affect gameplay of existing main roster characters in any way: basic play, targeted lines, and epilogue select all work correctly. However, if you encounter any problems related to this, we can consider rolling back this change until a fix can be found.

As a test, I have added a new group: Offline Opponents #1. This should work in your local copy but fail to load on the spnati.net version. Let me know if we get unexpected results.