Wii Fit Trainer 3.0
UnderscorM3 has undertaken a dialogue update for the Wii Fit Trainer, doubling her lines from about 200 to about 400!

As this is a very significant update that is not bug-free, I have moved the Wii Fit Trainer on to the testing tables ahead of schedule. She now has skips the projected break from playability. Please test her out and give feedback!

Note that she currently has several unassigned images that will appear broken. A list of these is in the Validator report and will be relayed to UnderscorM3.