Kyu Changes (Manual)
Doing this manually.

Notes from YSP:

Alters some poses.
Began to add the groundwork for more circumstantial lines (Kyu will soon be receiving more lines for if the player loses, if the player loses incredibly badly, and if she and the player are the last two in the game).
Adds a couple of targeted lines for DMG, Steve, and Futaba.