Elaine's Big Update
This is an update from /u/sjrslev.

This update is a huge graphical update as well as dialogue update.

I've tested Elaine once and made a few spelling corrections to her .xml using the editor, but overall she was pretty bug-free!

A word of warning, this update increases Elaine's stages by 1, which will break many lines targeting her.

Pit and Meia in particular may need to make adjustments following this.

[This also includes a few bugfixes to Marinette's dialogue, but doesn't add anything new.]