Added D.Va's side bet
If D.Va gets a good (or okay) hand and has lost no clothing in the sixth round, she'll propose a side bet that the first loser does anything the winner asks.

There is the potential for more targeted lines here when other characters go out first, but I included a couple of samples to start with.

As always, reading the source text file gives a quicker overview of the changes.