Allow name, "size", and clothing selection to be saved separately for male and female player character.
Done by storing those attributes in subobjects of the save object and loading the other subobject when switching gender.

The mergeObjects function is made recursive in case a part of a subobject is missing for some reason.

Only drawback is that an invalid selection (more than 8 items of clothing) can become saved in the cookie.