Make auto forfeit work consistently (in particular, it was straight up broken in the "Wait" state).
I did this by having the functions that can either set up a timer or complete their work straight away (including tickForfeitTimers()) return true in the former case and false if advanceGame() should go ahead and set up the auto forfeit timed if enabled and the human player is out, or otherwise enable the main button.

However, in those cases were special timers are used (cards dealt, AIs done exchanging cards, after someone's orgasm, and after the human has chosen a piece of clothing to strip), the same thing (enabling the main button or setting up the auto forfeit timer) needs to happen so I created a function allowProgression() that does that. It might easier to understand to the code if it's simply called wherever needed instead of returning true or false.

Also got rid of the savedContext global variable, instead passing the value as another argument to finishMasturbation().

Also, "Cumming..." is not a context that can happen since tickForfeitTimers() returns null when it changes the main button caption to that.

Also moved "Next" state to it's place in the ordering.