Incremented make_xml.py's version number to 1.49
make_xml.py underwent changes in the most recent commit, and a new version number imprinted on the files it makes will help with troubleshooting and could help us iin the future to identify who is still using an older version.

Calling it 1.49 will differentiate it from the *previous* version 1.48, which was buggy.