Added a few targeted lines from Nagisa to Amalia
More targeted lines will be needed toward Amalia later because she removes her upper and lower important layers twice each, and I don't want Nagisa to act like it's the first time when it's the second. But for now, this is a start.

I also adjusted a few hand quality cases based on the recent reclassification of Two Pair to being a good_hand.