Added some markers to Sayla
I don't know if this is technically in the scope of my permissions anymore, but I used to do little repair tweaks to inactive characters all the time back in the old days. If anyone was working on a Sayla update that conflcts with this one, please feel free to bulldoze this. I'll work out where to restore the markers when the dust settles.

The adjustments are small and won't be noticed by players:
- Added markers to a few of Sayla strip lines
- Corrected ~name~ to ~Name~ at the start of sentences, which didn't exist when Sayla was created
- Added her new tags to her text document, except "sayla" which should now be automatic
- Re-ran make_xml, which restored a couple of alternative start lines that already existed