Refactor poker code sililarly to v12
Define Card and Hand classes; make ranks and suits summary arrays members of Hand so that they can be reused by AI.

As in v12, Ace is represented as 14 instead of 1 so that cards to be exchanged can be found by comparing them to he value array elements set by the determine() method of Hand.

A simpler algorithm for finding straight draws is included. It also fixes the bug that traight draws are not attempted if the odd card out is higher than the cards to keep.

More OOfication remains to be done.