Added the "selected" line to make_xml.py
With MR !674 (for issue #19), ReformCopyright gave us the exciting ability to use our flexible toys in pre-game lines and introduced the concept of the "selected" state, which is when a character has been selected but the game hasn't started. This is formerly when a character would have drawn a second start line, and this behaviour continues for most characters.

I have clumsily added "selected" line functionality to make_xml.py. Please advise anyone anywhere if you encounter any issues with this.