Incremented make_xml.py version number stamping to 1.54
ReformCopyright removed the mandatory requirement of having a `start` file in each text file, as these have been replaced by `selected` and `game_start` lines. Future xmls will be stamped "1.54" to help us troubleshoot this issue for people in the coming months.

Watch this space for a future solution by RC to the "Your pants is nice" problem, which will involve indicating which of your characters' clothing items sound plural.