Incremented make_xml.py version number stamping to 1.55
ReformCopyright is working on allowing writers to vary a line depending on whether a piece of clothing sounds singular or plural. However, this requires that every character be updated with an extra variable for each piece of clothing. This is not yet supported in the Character Editor, so do not start to write these lines yet.

If you use make_xml, you can now specify a plural-sounding item of clothing as follows:

`Shoes,shoes,extra,other,plural`