make_xml will now correct ~name~ to ~Name~ at the start of sentences
`~Name~` is `~name~`, but with the first letter capitalized. There are a few places you can use this, but the most obvious is at the start of a sentence. This will only be relevant if the opponent's label is lower case, which is currently most likely if it is the player's name. Likewise, Jasmine's initial label should start with a lower case letter. (#38 proposes other lower case label possibilities.)

This commit helpfully changes any `~name~` it sees at the start of a line into a '~Name~`. If you start any sentences mid-line, be aware that these will continue to not be auto-fixed.

I have increased the make_xml version number to 1.56. Though this is a tiny change, bumping up the version number with each alteration will make for easier troubleshooting when helping people with outdated versions in the months to come.