Added plural values to Jasmine's clothing
To support forthcoming line functionality that can assess whether an item of clothing sounds plural, each character will need to be updated with a `plural=true` on selected lines in their wardrobe list. This is supported by the latest versions of the Character Editor (from v2.1.1) and make_xml (from v1.55).

I'm going through and updating the xml files of characters who have had no content edits in the past six months. If you see the name of character you're working on, watch out for edit conflicts.

In addition to the clothing plural changes, I have changed Jasmine's initial label from "Anonymous girl" to "anonymous girl. Because this is label called with `~name~`, the capital letter was showing up mid-sentence. To prevent the lower case letter to Jasmine (and any other opponent or player) from being at the start of a sentence, you should start sentences with `~Name~`, which is the same but with a capital letter at the beginning. This can be done quickly by opening your xml in a simple text editor and doing a find-replace of `">~name~` to `">~Name~`. (This won't catch mid-line `~name~`s, though.)