Hashko
This is a very early version of "Hashko". And by "very early", I mean that work begun on her in November 2017 and has progressed *very slowly* since then. This is not a complete product, but she's ready to be exposed to the Testing Tables for some early feedback. She'll be repetitive, and she doesn't have enough targeted lines for a full release, but she has the seed of what will mature into a full character. I'm very aware of some of her failings; it's the failings that I'm not aware of that will prove most useful to learn about.

With this initial commit, "Hashko" (Sanako Isogai) has 518 unique lines.