Increased Sanako's AI slightly
518 -> 530 unique lines

Sanako is a 5-layer character, meaning she has 6 stages that need difficulty assigned. She started life as an all-average character. However, this is likely to make her reach the end of her layers quicker than an average 8-layer oppnent. I have adjusted Sanako's AI as follows:

* good
* average
* average
* average
* good
* good

So now she's "good" for one layer to crack her outer shell, and she has two more "good" layers when she's down to her underwear. I will monitor this in test games to determine if she's holding on too long. If she is, I might move to a 4:2 balance between average and good. This is not an effort to make her harder; it's just no good if she loses before anyone else is exposed.