Make all substitution variables case insensitive, except that if the first letter is uppercase, the result gets its first letter uppercased, and stop using the "proper" clothing names.
behaviour_template.txt says that the first field of a clothes line is the "formal name", but in reality it's had to be the same as the "lower case name" with the initial letter in uppercase, because it was mapped to ~Clothing~, which is used in the beginning of sentences.

With this change, that field can really be used for a more colorful description, but we'll have to allow some time for make_xml.py users to be informed that they need to update their clothes lines.

Also, preparations are made to rename the "lowercase" attribute "genericName", so as not to suggest that the other attribute should also be lowercase.