Incremented make_xml.py version number stamping to 1.58
As the release number is now handled by listing.xml, make_xml no longer needs to output it to the behaviour.xml file.

Note: If you are using make_xml on text files that do not have all cases images filled in, even on blank lines, version 1.56 is still the last usable version as per Issue !40.

https://gitlab.com/spnati/spnati.gitlab.io/issues/40