Removes Sei's epilogue badge
We've been getting a lot of bug reports and complaints that people are not getting an epilogue for Sei, even though she does have an epilogue badge. This is because her epilogues has requirements that a majority of players do not meet in casual play (either play as a female, or play with Meia+Sei together). We decided to remove her epilogue badge temporarily, until we've implemented a proper feature to indicate the requirements needed to unlock conditional epilogues.