Incremented make_xml.py version number stamping to 1.59
ReformCopyright has made a few changes recently to make_xml that warrant a version number increase. This number is stamped on to each file made with this script, and this version stamping can help us troubleshoot future problems by identifying which version of the script the behaviour file was created with.

Changes:
* Implimented "direction" and "location" options for speech bubble arrows. (usage not documented)
* Changed how ~silent~ works in the game, and therefore how it works inside make_xml. (usage unchanged)
* Changed hand type names to no longer be case sensitive. (usage unchanged)