Incremented make_xml.py version number stamping to 1.60
ReformCopyright has added additional ways to check the state of the game in MR !867. These are complex and will probably need some sort of documentation before we can understand them.

As part of this update, RC adjusted make_xml to be able to use these features. To help troubleshoot issues experienced by people currently living in the future, I have increased make_xml's date stamping to indicate that the most recent behaviour.xmls will be created with this improved version.