Compressed Nugi-chan's images
I compressed all of Nugi-chan's images, including her epilogue, reducing her total image file size from 19.7 to 10.8 megabytes. This is a saving of 45%.

All compression was done via TinyPNG.com. Note that this is lossy compression. You can see the affect it had on gradients by comparing the gray curtains in Nugi-chan's first epilogue image. (To do this in the GitHub Desktop Client, click the History tab.) However, as the differences are ultimately minimal, these kinds of savings are worthwhile and important as we continue to endeavor to fit more characters into the game.