Load player tag choices from cookie (also save them differently).
To make loading tag choices simpler, they are saved as an object, and only a singular value is saved per choice, with implicit tags added when proceeding to the game screen.

Also make the form generation code easier to read with the help of DOM objects and jQuery.

TODO: Save the precise input skin color slider value instead of the tag.