Let already selected characters greet subsequently selected opponents.
Note: Only when individuallty selected; when a group is selected, all characters arrive simultaneously, so it doesn't make sense to trigger an opponent_selected event, and all slots are prepared so they can all see which other characters are selected.

When using the random buttons, up to three opponents could already be present, but it's still to complicated to handle.

I hope I sprinkled the right amount of volatile behaviour updating and behaviour committing.