Avoid creating preset tables where not all characters are available.
This changes the strategy from trying to load metadata for all characters included on preset tables
(and failing on spnati.net when it comes to the offline tables) to only setting up those tables where
all characters' development statuses are among those enabled, or "testing". All characters in testing
should be on a testing table, so we don't load anyone needlessly.