Add opponent_stripping and opponent_stripped in the same vein as opponent_lost.
Can (should?) be used when the target is a specific character and stage anyway, so that any change to the type of clothing doesn't break the targeting. Can also be used when the type of clothing (and target gender) doesn't matter because we're targeting the removal of the last layer, for example.