Added a few targeted lines to Nugi-chan
587 -> 598 unique lines

Here's a few targeted lines and a few overdue fixes for Nugi-chan. I should convert to the Character Editor because writing targeted lines in the text document is such a pain to research, but I like to vary my poses per stage for the one targeted line, so for now I am trapped. :(

Also replaced Nugi-chan's <3 heart with a ♡ heart because that's how the kids do it these days. (Not that Nugi-chan isn't an adult. According to her creator, she is exactly eighteen., which is arguably both a kid and an adult. So I guess she was born *after* The Matrix was released in 1999. Weird.)