Fix forced alternate costume issues:
* Use selection_image in opponent selection quads.
* Keep costume selection. Only unload selected alternate costume if another costume is selected (including default).
* Apply forced selection by calling .selectAlternateCostume() when Opponent object is instantiated instead of doing that when loading behaviour.xml and only setting selection_image upon instantiation.
* Initialize costume members in Opponent() rather than Player() since Opponent() doesn't call Player().

Note that selected_costume is the selection on the selection screen (the costume folder name) and alt_costume is the object containing metadata of the loaded costume. We need to add the costume base folder to that object to be able to know if the player has selected a different costume.