Add targetSaying and alsoPlayingSaying, volatile conditions that look at the (raw) dialogue.
The conditions are fulfilled if the given string is contained within the raw dialogue of the target or also playing character (matched case insensitively).

It looks at the raw dialogue (for now) because the dialogue only expanded after a state has been committed. Though it's possible that we could expand the raw dialogue when testing these conditions, that might make these conditions harder to use.