Added Nagisa's first interview response to Streaming-chan
The real change here is that Nagisa's behaviour.xml is finally adapted to use the Character Editor! FINALLY! Because the CE doesn't handle out-of-stage images that start with numbers very well, I've renamed "7-rear.png" to "seven-rear.png".

The feature that finally won me over was the visbility and easy of using alsoPlayingSayingMarker and position detection. Nagisa has a very early and VERY limited trial of this in her first response to Streaming-chan's interview questions. Specifically, if Streaming-chan is saying a marked line AND is to Nagisa's left, Nagisa will respond immediately in the same phase. She will likewise respond to the next two things SC says, filling in her half of the conversation. However, if Streaming-chan is to the right, Nagisa will respond in the NEXT phase. She actually responds slightly differently because the context of the question is gone.

Right now, Nagisa only says this for one (random) question and only at stage zero, but this could revolutionize interactivity if copied and improved. See Nagisa's channel for more details.